refactor: rename resizer components to be more generic
