+++ ui/src/notebooks/pipes/Query/ResizerHeader.tsx
